vdb/download.py: Let --fasta_fields take multiple values
Otherwise it's not possible to specify a useful set of fields on the
command line.  This hasn't been an issue since currently all the
vdb/*_download.py scripts overwrite any passed value with a hardcoded
list of fields.  For a pathogen build system, however, it's useful to
specify these on the command line.