Fix imports from parent modules in vdb and tdb for python 3
The original fix for importing from parent modules in python 2 did not work for
python 3. The ideal solution would be to reorganize fauna to avoid this type of
import, but this commit implements a short-term solution that works in both
python 2 and 3 by prepending the fauna directory to the python path.