tdb/cdc_upload: Add `-rename` option
The CDC is moving to a new table format that has different column names.
The new `--rename` option prints the titers to a new TSV with columns
that have been renamed according to `source-data/cdc_titer_column_map.tsv`.
The column names should match previous CDC titer table columns so that
the `upload` and `parse` classes do not need to be updated.