cdc_titer_column_map: set sr_lot to serum_id
The CDC uses the lot numbers in the `sr_lot` column to discuss titer
measurements, so it makes more sense to use the lot numbers as the
`serum_id`.

This commit also removes the mapping for `sr_ferret` to `ferret_id`
because the `read_flat` method within `tdb/parse` assigns this
`ferret_id` to the `serum_id`, which would overwrite the lot number.
Since we don't need the ferret id as the serum id anymore, there's no
reason to include it in the mapping.

Resolves #126.