cdc_upload: add `get_serum_host`
Returns the serum host of a measurement based on whether the host
string is present in the measurement's serum_id (case insensitive).
Currently only searches for "human" and "mouse" because these are known
serum hosts that we've seen in the past CDC titer data.
If there are no matches, then returns "ferret" by default.

Resolves #130