flu_upload: convert all nan values to None
The dtype of the column affects where the nan -> None conversion happens.
Convert all dtypes to "object" so that all nan values are converted
based on https://stackoverflow.com/a/70803926. The dtype will be changed
to object anyway once a `None` value is included in the column.

This change was needed when I was re-uploading Yam sequences that was
mislabelled as Vic sequences.¹ Because these are older records (~1990s),
they had completely empty columns that were read in as float columns,
which failed the previous `.where` conversion.

¹ https://bedfordlab.slack.com/archives/C03KWDET9/p1692722299053159