delete_records: Add `--interval` option
Add a new `--interval` option to filter records with date intervals.
This will be used to delete records that uploaded within a certain
date interval in preparation for uploading the latest CDC data¹ and
for merging https://github.com/nextstrain/fauna/pull/131.

Follows the pattern of the `--interval` option for `tdb/download.py` and
uses the `rethinkdb_date_greater` function from `vdb/download.py` to
check for date range in documents.

¹ https://bedfordlab.slack.com/archives/C03KWDET9/p1697750132854499