tdb/elife_upload: Update `parse_assay_date_from_filename`
Use regex to find all matches for the expected date format 'YYYYMMDD'.
Then use the datetime module to validate the date string and check that
the date is earlier than the date we are parsing the file.

I made the decision to use the latest date if there are multiple matches
with the expectation that we would be parsing files not long after the
assay date.