Update zika instructions
The old instructions were written for ViPR which became obsolete and was replaced
by BV-BRC. The old instructions no longer work and we have since moved to using
NCBI datasets for downloading sequences and metadata files.

The filtering steps are already part of the phylogenetic build steps so are
no longer a consideration during ingest. Point team members to how to ingest
recent zika data and push to the nextstrain data endpoint. Point team members
to the current phylogenetic build steps.