vdb/flu_upload: only prefix accession with "EPI" if needed
Digging through git history and found example of FASTA header that
suggests GISAID did not used to include the "EPI" prefix in their
DNA Accession no. field.¹

This must have changed around October 2023 because we have sequences
with "EPIEPI" accessions for sequences submitted after September 27th 2023.²

This commit changes the ingest to only prefix with "EPI" if the accession
does not already have the prefix.

¹ https://github.com/nextstrain/fauna/blame/f485baa3621002b3ff6f833c743180239a92bf14/vdb/gisaid_flu_upload.py#L281-L282
² https://bedfordlab.slack.com/archives/C03KWDET9/p1700609695217959