script/delete_records: Add option to match fields with regex pattern
Uses rethinkdb's `match` command to filter for records with field value
that matches the provided regex pattern. See rethinkdb docs for more
details: https://rethinkdb.com/api/python/match/

This was prompted by our need to delete flu sequence records that have
accessions with pattern "EPIEPI". We've fixed the accession with #148,
but we need to manually remove the old duplicate sequence records
because the flu sequence table uses the accession as the index.¹

¹ https://github.com/nextstrain/fauna/blob/ec1feb679715890ae6d14efe11c979f27d6f1d6f/vdb/upload.py#L82