Update the zika ingest and build instructions (#147)
The old instructions were written for ViPR which became obsolete and was replaced
by BV-BRC. The old instructions no longer work and we have since moved to using
NCBI datasets for downloading sequences and metadata files.

The filtering steps are already part of the phylogenetic build steps so are
no longer a consideration during ingest. 

The purpose of the page is to provide a soft on-boarding for a specific use case: 
allowing an internal person to update our public builds. Therefore, the instructions 
document the current status quo for pulling the appropriate branch, ingesting the data, 
uploading that data, and updating the build. Each section provides commands they can 
copy and paste, while also linking to the official READMEs for more detail.

These instructions may be updated if and when the production builds are merged to the default branch.

* fixup: github markdown note
Co-authored-by: Victor Lin <13424970+victorlin@users.noreply.github.com>