vdb/flu_upload: Exclude bad submitting lab names
In testing, I found the `--overwrite` flag does not completely replace
the old records with the new ones. It does a merge of the two records,
giving preference to values in the new record.

So, in order to drop the "$ins_submitting_name" values in the
`flu_sequences` table, we must include the `submitting_lab` field from
the FASTA header, but replace the bad submitting lab with an empty value.