vidrl_upload/read_flat_vidrl: standardized human sera name and ids
Strip the "pool" suffix from the serum strain name and standardized the
serum id.

While looking into this change, I discovered that the strain name used
for the human sera references in H1 and H3 is the egg vaccine strain
regardless of passage annotation. Currently unclear if this is an error
in the flat files or if we've misunderstood the passage annotations for
human sera data. Once we clear this up, we should add some type of
vaccine strain verification so that we can flag mismatches like this
automatically.