read_flat_vidrl: Validate records
Validate records in single flat file. Ensure that the serum
abbreviations map to a single serum strain and all records have the
same test date.

As a side effect, the validated `serum_abbr_map` and `test_date` are
returned to be used for processing the reference panel records in
following commits.