read_flat_vidrl: ingest reference_panel.csv file
Ingests the matching "*_reference_panel.csv" for a provided
"*_flat_file" fstem if the reference panel file exists. The records
parsed from the reference panel file is appended to the same tmp
file that is then passed to elife_upload.py.

This currently includes "extra" records in comparison to Excel files,
where the human sera pool strain is the "test virus" against the other
references. If I strip the `pool` suffix from human sera pool strain,
the measurements are exact duplicates of the measurements for the
matching reference strain. We will need to decide whether or not these
records should be dropped.