vdb/download: Add `--prioritized_seqs_file` option
Allow users to define prioritized sequence accessions for a strain
when the duplicate resolve method is `split_passage`. This flag could be
extended for other resolve methods, but we are not actively using them
in seasonal flu workflows so punting for now.

The file is expected to be a TSV file with a `strain` and `accession`
column where the `strain` column only contains unique strain names.
With the seasonal flu's pattern of downloading each segment separately,¹
the file should contain sequence accessions for a single segment.

¹ <https://github.com/nextstrain/seasonal-flu/blob/73b8294906da99a6abd26f543edd68dfd6e63d0e/workflow/snakemake_rules/download_from_fauna.smk#L69>