tdb/cdc_upload: Add `--curate-output` option
Outputs the curated measurements to specified NDJSON file without
uploading to the database. This includes the curations that are done
in tdb/download.py. I figured it's cleaner to output to an NDJSON file
instead of outputting to stdout because there are so many log/print
statements throughout the code that would intermix with the records.

We now have 3 options:

1. `--curate-output` to output all curated records to NDJSON
2. `--preview` to preview one example record
3. excluding both options will upload records to the database